main layout

+++ MainCore/Commands/UI/MainLayout/DeleteAccountCommand.cs
@@ -2,15 +2,15 @@
 using MainCore.Entities;
 using MainCore.Notification.Message;
 using MainCore.Repositories;
+using MainCore.Services;
+using MainCore.UI.ViewModels.UserControls;
 using MediatR;
-using ReactiveUI;
-using System.Reactive.Linq;
 
 namespace MainCore.Commands.UI.MainLayout
 {
-    public class DeleteAccountCommand : ByAccountIdBase, IRequest
+    public class DeleteAccountCommand : ByListBoxItemBase, IRequest
     {
-        public DeleteAccountCommand(AccountId accountId) : base(accountId)
+        public DeleteAccountCommand(ListBoxItemViewModel items) : base(items)
         {
         }
     }
@@ -19,20 +19,30 @@ namespace MainCore.Commands.UI.MainLayout
     {
         private readonly IMediator _mediator;
         private readonly IUnitOfRepository _unitOfRepository;
+        private readonly IDialogService _dialogService;
 
-        public DeleteAccountCommandHandler(IMediator mediator, IUnitOfRepository unitOfRepository)
+        public DeleteAccountCommandHandler(IMediator mediator, IUnitOfRepository unitOfRepository, IDialogService dialogService)
         {
             _mediator = mediator;
             _unitOfRepository = unitOfRepository;
+            _dialogService = dialogService;
         }
 
         public async Task Handle(DeleteAccountCommand request, CancellationToken cancellationToken)
         {
-            var accountId = request.AccountId;
-            await Observable.Start(() =>
+            var accounts = request.Items;
+            if (!accounts.IsSelected)
             {
-                _unitOfRepository.AccountRepository.Delete(accountId);
-            }, RxApp.TaskpoolScheduler);
+                _dialogService.ShowMessageBox("Warning", "No account selected");
+                return;
+            }
+
+            var result = _dialogService.ShowConfirmBox("Information", $"Are you sure want to delete \n {accounts.SelectedItem.Content}");
+            if (!result) return;
+
+            var accountId = new AccountId(accounts.SelectedItemId);
+
+            _unitOfRepository.AccountRepository.Delete(accountId);
             await _mediator.Publish(new AccountUpdated(), cancellationToken);
         }
     }
